fix: put backend host in proxy request headers log
The issue was due to a reference problem. We are snapshotting the
request before calling the backend. However, in our case, the connector
is updating the `Host` header, and we didn't have this change in the log
object.

To fix that, we update the proxy request log object with the request
header in the post step to have all the headers, even those added by the
connector.

https://gravitee.atlassian.net/browse/APIM-1157
