+++ gravitee-apim-gateway/gravitee-apim-gateway-tests-integration/src/test/java/io/gravitee/gateway/tests/http/logging/LoggingIntegrationTest.java
@@ -28,17 +28,17 @@ import io.gravitee.apim.gateway.tests.sdk.annotations.GatewayTest;
 import io.gravitee.apim.gateway.tests.sdk.configuration.GatewayConfigurationBuilder;
 import io.gravitee.apim.gateway.tests.sdk.reporter.FakeReporter;
 import io.gravitee.common.http.HttpStatusCode;
-import io.gravitee.definition.model.Api;
-import io.gravitee.definition.model.Logging;
-import io.gravitee.definition.model.LoggingContent;
-import io.gravitee.definition.model.LoggingMode;
-import io.gravitee.definition.model.LoggingScope;
+import io.gravitee.definition.model.*;
 import io.gravitee.gateway.reactor.ReactableApi;
 import io.gravitee.reporter.api.log.Log;
+import io.reactivex.rxjava3.subjects.BehaviorSubject;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
+import io.vertx.junit5.VertxTestContext;
 import io.vertx.rxjava3.core.http.HttpClient;
 import java.util.concurrent.atomic.AtomicReference;
+import org.assertj.core.api.SoftAssertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -52,10 +52,30 @@ import org.junit.jupiter.api.Test;
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class LoggingIntegrationTest extends AbstractGatewayTest {
 
+    BehaviorSubject<Log> subject;
+
+    @BeforeEach
+    void setUp() {
+        subject = BehaviorSubject.create();
+
+        FakeReporter fakeReporter = getBean(FakeReporter.class);
+        fakeReporter.setReportableHandler(reportable -> {
+            if (reportable instanceof Log) {
+                subject.onNext((Log) reportable);
+            }
+        });
+    }
+
     @Override
     protected void configureGateway(GatewayConfigurationBuilder gatewayConfigurationBuilder) {
         super.configureGateway(gatewayConfigurationBuilder);
         gatewayConfigurationBuilder.set("api.jupiterMode.enabled", "true");
+    }
+
+    @Override
+    public void configureApi(Api api) {
+        super.configureApi(api);
+        api.setExecutionMode(ExecutionMode.JUPITER);
     }
 
     @Override
@@ -65,31 +85,85 @@ class LoggingIntegrationTest extends AbstractGatewayTest {
         logging.setContent(LoggingContent.HEADERS_PAYLOADS);
         logging.setScope(LoggingScope.REQUEST_RESPONSE);
         logging.setCondition("{#response.status == 200}");
+
         if (api.getDefinition() instanceof Api) {
             ((Api) api.getDefinition()).getProxy().setLogging(logging);
         }
     }
 
     @Test
-    void should_log_everything(HttpClient httpClient) throws Exception {
+    void should_log_everything(HttpClient httpClient, VertxTestContext context) throws Exception {
         JsonObject mockResponseBody = new JsonObject().put("response", "body");
         wiremock.stubFor(get("/endpoint").willReturn(okJson(mockResponseBody.toString())));
 
-        FakeReporter fakeReporter = getBean(FakeReporter.class);
-        AtomicReference<Log> logRef = new AtomicReference<>();
-        fakeReporter.setReportableHandler(reportable -> {
-            if (reportable instanceof Log) {
-                logRef.set((Log) reportable);
-            }
-        });
-
         AtomicReference<String> requestHostAndPortRef = new AtomicReference<>();
-
         JsonObject requestBody = new JsonObject().put("field", "of the pelennor");
+
+        subject
+            .doOnNext(log -> {
+                final String transactionAndRequestId = wiremock
+                    .getAllServeEvents()
+                    .get(0)
+                    .getRequest()
+                    .getHeaders()
+                    .getHeader("X-Gravitee-Transaction-Id")
+                    .firstValue();
+
+                SoftAssertions.assertSoftly(soft -> {
+                    soft.assertThat(log.getClientRequest().getBody()).isEqualTo(requestBody.toString());
+                    soft.assertThat(log.getClientRequest().getUri()).isEqualTo("/test");
+                    soft
+                        .assertThat(log.getClientRequest().getHeaders().toSingleValueMap())
+                        .contains(
+                            entry("host", requestHostAndPortRef.get()),
+                            entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
+                            entry("X-Gravitee-Request-Id", transactionAndRequestId),
+                            entry("content-length", String.valueOf(requestBody.toString().length()))
+                        );
+                });
+
+                SoftAssertions.assertSoftly(soft -> {
+                    soft.assertThat(log.getProxyRequest().getBody()).isEqualTo(requestBody.toString());
+                    soft.assertThat(log.getProxyRequest().getUri()).isEqualTo("http://localhost:" + wiremock.port() + "/endpoint");
+                    soft
+                        .assertThat(log.getProxyRequest().getHeaders().toSingleValueMap())
+                        .contains(
+                            entry("Host", "localhost:" + wiremock.port()),
+                            entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
+                            entry("X-Gravitee-Request-Id", transactionAndRequestId),
+                            entry("content-length", String.valueOf(requestBody.toString().length()))
+                        );
+                });
+
+                SoftAssertions.assertSoftly(soft -> {
+                    soft.assertThat(log.getProxyResponse().getBody()).isEqualTo(mockResponseBody.toString());
+                    soft.assertThat(log.getProxyResponse().getStatus()).isEqualTo(200);
+                    soft
+                        .assertThat(log.getProxyResponse().getHeaders().toSingleValueMap())
+                        .doesNotContainKeys("X-Gravitee-Transaction-Id", "X-Gravitee-Request-Id")
+                        .contains(entry("Content-Type", "application/json"));
+                });
+
+                SoftAssertions.assertSoftly(soft -> {
+                    soft.assertThat(log.getClientResponse().getBody()).isEqualTo(mockResponseBody.toString());
+                    soft.assertThat(log.getClientResponse().getStatus()).isEqualTo(200);
+                    soft
+                        .assertThat(log.getClientResponse().getHeaders().toSingleValueMap())
+                        .contains(
+                            entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
+                            entry("X-Gravitee-Request-Id", transactionAndRequestId),
+                            entry("Content-Type", "application/json")
+                        );
+                });
+            })
+            .doOnNext(m -> context.completeNow())
+            .doOnError(context::failNow)
+            .subscribe();
 
         httpClient
             .rxRequest(HttpMethod.GET, "/test")
             .flatMap(request -> {
+                request.setHost("127.0.0.1");
                 requestHostAndPortRef.set(request.getHost() + ":" + request.getPort());
                 return request.rxSend(requestBody.toString());
             })
@@ -106,48 +180,5 @@ class LoggingIntegrationTest extends AbstractGatewayTest {
             });
 
         wiremock.verify(getRequestedFor(urlPathEqualTo("/endpoint")));
-        final Log log = logRef.get();
-        final String transactionAndRequestId = wiremock
-            .getAllServeEvents()
-            .get(0)
-            .getRequest()
-            .getHeaders()
-            .getHeader("X-Gravitee-Transaction-Id")
-            .firstValue();
-
-        assertThat(log.getClientRequest().getBody()).isEqualTo(requestBody.toString());
-        assertThat(log.getClientRequest().getUri()).isEqualTo("/test");
-        assertThat(log.getClientRequest().getHeaders().toSingleValueMap())
-            .contains(
-                entry("host", requestHostAndPortRef.get()),
-                entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
-                entry("X-Gravitee-Request-Id", transactionAndRequestId),
-                entry("content-length", String.valueOf(requestBody.toString().length()))
-            );
-
-        assertThat(log.getProxyRequest().getBody()).isEqualTo(requestBody.toString());
-        assertThat(log.getProxyRequest().getUri()).isEqualTo("http://localhost:" + wiremock.port() + "/endpoint");
-        assertThat(log.getProxyRequest().getHeaders().toSingleValueMap())
-            .contains(
-                entry("Host", "localhost:" + wiremock.port()),
-                entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
-                entry("X-Gravitee-Request-Id", transactionAndRequestId),
-                entry("content-length", String.valueOf(requestBody.toString().length()))
-            );
-
-        assertThat(log.getProxyResponse().getBody()).isEqualTo(mockResponseBody.toString());
-        assertThat(log.getProxyResponse().getStatus()).isEqualTo(200);
-        assertThat(log.getProxyResponse().getHeaders().toSingleValueMap())
-            .doesNotContainKeys("X-Gravitee-Transaction-Id", "X-Gravitee-Request-Id")
-            .contains(entry("Content-Type", "application/json"));
-
-        assertThat(log.getClientResponse().getBody()).isEqualTo(mockResponseBody.toString());
-        assertThat(log.getClientResponse().getStatus()).isEqualTo(200);
-        assertThat(log.getClientResponse().getHeaders().toSingleValueMap())
-            .contains(
-                entry("X-Gravitee-Transaction-Id", transactionAndRequestId),
-                entry("X-Gravitee-Request-Id", transactionAndRequestId),
-                entry("Content-Type", "application/json")
-            );
     }
 }
